Fix generation of NuGet package
The nuspec file should not contain references to files that do not exist. If building in Windows, for example, only the Windows sealc.dll binary should be present in the nuspec file.
Also, we need to include the documentation file (SEALNet.xml) so that users can have IntelliSense when using SEAL.